requirements.txt
flask==2.2.2
flask_jwt_extended==4.4.4
flask-restful==0.3.9
requests==2.28.1
celery==5.2.7
redis==4.3.4
PyJWT~=2.5.0
